catalog_app/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4
drf-yasg==1.17.1
